python/requirements-dev.txt
llangchain-core
langchain
langchain-chroma>=0.1.2
pypdf
transformers
accelerate
chromadb
torch
gpt4all
